refactor: split telegram handlers
